LT: fix hanging because of ushort overflow
MSM2 has 92k entries, and uses 32 bit ints for indexes in 0CD2CFE9 section

+++ DAT1/Helpers/LocalizationHelper.cs
@@ -100,9 +100,9 @@ namespace DAT1.Files {
 			l.SortedKeyHashesSection.Values.Clear();
 			l.SortedIndexesSection.Values.Clear();
 
-			List<(uint, ushort)> hashesWithIndexes = new();
-			for (ushort i = 0; i < n; ++i) {
-				hashesWithIndexes.Add((l.KeyHashesSection.Values[i], i));
+			List<(uint, uint)> hashesWithIndexes = new();
+			for (uint i = 0; i < n; ++i) {
+				hashesWithIndexes.Add((l.KeyHashesSection.Values[(int)i], i));
 			}
 			var compare = (uint a, uint b) => {
 				if (a == b) return 0;
